requirements.txt
python-dotenv==1.0.1     # load env locally (optional)
pandas==2.2.2            # data
numpy==1.26.4            # math
requests==2.32.3         # HTTP (Finnhub)
plotly==5.24.0           # charts
streamlit==1.38.0        # dashboard
google-cloud-storage==2.18.2  # GCS read/write
functions-framework==3.*
yfinance